+++ src/tasks/airquality.py
@@ -54,8 +54,9 @@ class AirQualityTask(object):
                 if (has_category_changed and is_current_within_threshold_of_prior):
                     self._tweetAirQuality(prior_air_quality, current_primary)
 
-            # Always save the current primary observation
-            self._saveAirQuality(current_primary)
+            # Save the current primary observation
+            if (prior_air_quality is None or current_primary.timestamp > prior_air_quality.timestamp):
+                self._saveAirQuality(current_primary)
 
             # Go to sleep for a little while
             self._sleep()