+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+### Enhanced
+- Only save the air quality observation when a new one is received
+
 ## Release 1.3.2 [2023-10-01]
 ### Fixed
 - Bump `airnowpy` from 2.2.1 to 2.2.2